+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+#### 🐞 Fixes
+
+- Fixed an issue where the new action pipeline wouldn't wait for in-flight moonbase (remote caching)
+  requests to finish. The old pipeline was not affected.
+
 ## 1.27.2
 
 #### 🐞 Fixes